Check fx account in home menu and main menu
